Refactor workflow engine into focused modules
- Split workflow limits, transitions, state, planner records, and review loop into dedicated modules
- Rename workflow policy surface to `WorkflowRunPolicy`
- Update prompt composition and workflow tests to match the new structure
